修复 windows 32bits 编译警告

+++ umysql/src/Connection.cpp
@@ -611,7 +611,7 @@ void *Connection::handleResultPacket(int _fieldCount)
 
   // Read Field packets
 
-  UMTypeInfo *typeInfo = (UMTypeInfo *) alloca(fieldCount * sizeof(UMTypeInfo));
+  UMTypeInfo *typeInfo = (UMTypeInfo *) alloca((size_t)(fieldCount * sizeof(UMTypeInfo)));
   int iTypeInfo = 0;
 
   while (true)
